separe un componente
